+++ Assignment-1/test/item_test.dart
@@ -1,5 +1,5 @@
 import 'package:test/test.dart';
-import '../bin/item.dart';
+import '../lib/item.dart';
 
 main() {
   test('Some Test Case', () {});
